202306280145 - 총 이용자수 HttpSessionListener 코드 리팩토링 202306281620 - 총 이용자수 HttpSessionListener 코드 삭제 / MEMBER TABLE COUNT DB 연동 코드 리팩토링

+++ trakker/src/main/java/com/example/trakker/item/RatingDTO.java
@@ -6,9 +6,9 @@ import lombok.ToString;
 @Data
 @ToString
 public class RatingDTO {
-    private Long t_num; // 시퀀스
-    private Integer mem_num; // 회원 번호
-    private Integer review_num; // 리뷰 번호
+    private long t_num; // 시퀀스
+    private long mem_num; // 회원 번호
+    private long review_num; // 리뷰 번호
     private Double rating; // 별점
 
     private Double ratingAvg; // 별점 총합
